pgwire: get rid of conn.execCfg
The reference from a connection to an ExecutorConfig is no longer
needed.

Release note: None
