+++ pkg/server/authentication.go
@@ -265,7 +265,7 @@ func (s *authenticationServer) verifyPassword(
 	ctx context.Context, username string, password string,
 ) (bool, error) {
 	exists, hashedPassword, err := sql.GetUserHashedPassword(
-		ctx, s.server.execCfg, s.memMetrics, username,
+		ctx, s.server.execCfg.InternalExecutor, s.memMetrics, username,
 	)
 	if err != nil {
 		return false, err
